Clean up Volume definitions
Current state is that {ban,priority,whitelist}.txt are mounted as single
files. This dosn't allow to access core dumps that are created on run
time to be accessed at run time. It is not possible to mount a non
existent core file on container creation time.

If you want to access a core file on run time mount 'serverfiles/' and
disable Volumes for every single file.